requirements.txt
speechrecognition==3.8.1
vosk~=0.3.30
PyAudio~=0.2.11
gTTS~=2.2.2
playsound~=1.2.2
requests~=2.25.0
ipregistry~=3.0.0
PyYAML~=5.3.1
numpy==1.18.5
tensorflow~=2.2.0
nltk~=3.6.2
Flask~=2.0.1
wikipedia~=1.4.0